Isolate bindings of Telegram Bot API in anoter module.
