+++ timeflies/utils.py
@@ -3,13 +3,11 @@
 
 import logging
 import re
+
 from enum import Enum
-
-
 from pprint import pprint
-from requests import Session
 
-from .settings import API_URL, API_TOKEN
+from .telegram import send_message
 from .models import User
 
 
@@ -19,42 +17,6 @@ class Command(Enum):
 
 
 command_patterns = {c: re.compile('/%s' % c.name) for c in Command}
-
-
-def send_request(method=None, params=None, sess=None):
-    if not sess:
-        sess = Session()
-
-    url = API_URL.format(token=API_TOKEN, method=method)
-    r = sess.get(url, params=params)
-
-    if r.status_code != 200:
-        logging.error('request failed with status code %d', r.status_code)
-        return None
-
-    content_type = r.headers.get('Content-Type', '')
-
-    if not content_type.startswith('application/json'):
-        logging.error('wrong content-type: %s', content_type)
-        return None
-
-    try:
-        json = r.json()
-    except ValueError:
-        logging.error('invalid json: %s', r.text)
-        return None
-
-    return json
-
-
-def get_updates(offset=None, limit=100, timeout=60, sess=None):
-    params = dict(offset=offset, limit=limit, timeout=60)
-    return send_request('getUpdates', params, sess)
-
-
-def send_message(chat_id, text, sess=None):
-    params = dict(chat_id=chat_id, text=text)
-    return send_request('sendMessage', params, sess)
 
 
 def add_user(user_info, database):
